Binary Tree Zigzag Level Order Traversal
This is Problem No. 103 in LeetCode Binary Tree Zigzag Level Order Traversal and in this problem the root of a binary tree is given and we need to return the zigzag level order traversal of its nodes' values. (i.e., from left to right, then right to left for the next level and alternate between).